fix(di): missing injection definition on release variant
