+++ feature/debug-settings/src/release/kotlin/net/thunderbird/feature/debug/settings/inject/FeatureDebugSettingsModule.kt
@@ -1,6 +1,9 @@
 package net.thunderbird.feature.debug.settings.inject
 
+import net.thunderbird.feature.debug.settings.navigation.NoOpSecretDebugSettingsNavigation
+import net.thunderbird.feature.debug.settings.navigation.SecretDebugSettingsNavigation
 import org.koin.dsl.module
 
 val featureDebugSettingsModule = module {
+    single<SecretDebugSettingsNavigation> { NoOpSecretDebugSettingsNavigation }
 }
