+++ feature/debug-settings/src/release/kotlin/net/thunderbird/feature/debug/settings/navigation/NoOpSecretDebugSettingsNavigation.kt
@@ -1,0 +1,11 @@
+package net.thunderbird.feature.debug.settings.navigation
+
+import androidx.navigation.NavGraphBuilder
+
+object NoOpSecretDebugSettingsNavigation : SecretDebugSettingsNavigation {
+    override fun registerRoutes(
+        navGraphBuilder: NavGraphBuilder,
+        onBack: () -> Unit,
+        onFinish: (SecretDebugSettingsRoute) -> Unit,
+    ) = Unit
+}
